// lms_fe/app/models/enum.ts
export enum NatureOfAssets {
    SECURED = "SECURED",
    UNSECURED = "UNSECURED"
}

export enum RiskProfile {
    LOW = "LOW", 
    MEDIUM = "MEDIUM",
    HIGH = "HIGH"
}

export enum AmortizationPeriodType {
    YEAR = "YEAR",
    MONTH = "MONTH",
}

export enum InterestRateType {
    FIXED = "FIXED",
    VARIABLE = "VARIABLE"
}


export enum ContractStatus {
    ACTIVE = "ACTIVE",
    CLOSED = "CLOSED",
    PENDING = "PENDING"
}

export enum PaymentStatus {
    SUCCESS = "SUCCESS",
    FAIL = "FAIL",
    ON_HOLD = "ON_HOLD"
}

export enum FormField {
    REQUIRE = "require field"
}